requirements.txt
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
celery==5.3.0
redis==5.0.0
django-celery-beat==2.5.0
python-decouple==3.8
langgraph
langchain
langchain-anthropic
langchain-openai
openai
anthropic
pydantic
numpy
tiktoken
python-dotenv